+++ examples/src/tiledb_dense_read_async.cc
@@ -39,7 +39,6 @@
  */
 
 #include <tiledb.h>
-#include <cstdio>
 
 // Simply prints the input string to stdout
 void* print_upon_completion(void* s);
@@ -82,7 +81,7 @@ int main() {
 
   // Print cell values
   uint64_t result_num = buffer_sizes[0] / sizeof(int);
-  printf("result num: %llu\n\n", result_num);
+  printf("result num: %llu\n\n", (unsigned long long)result_num);
   printf(" a1\t    a2\t   (a3.first, a3.second)\n");
   printf("-----------------------------------------\n");
   for (uint64_t i = 0; i < result_num; ++i) {
